requirements.txt
tldextract==5.1.2
python-Levenshtein==0.25.1
